Revert "ACCUMULO-2589 Drop deprecated code from 1.x"
Restore all of the aggressively removed deprecated code, except the
trace module, which is no longer needed.

This reverts commit b69291a3453fd0e6091586cf37fb1636a356caa9.

Conflicts:
	minicluster/src/test/java/org/apache/accumulo/minicluster/impl/MiniAccumuloConfigImplTest.java
	server/base/src/test/java/org/apache/accumulo/server/util/FileUtilTest.java
	shell/src/main/java/org/apache/accumulo/shell/Shell.java

+++ core/src/main/java/org/apache/accumulo/core/client/impl/ThriftScanner.java
@@ -214,7 +214,7 @@ public class ThriftScanner {
     return (long) (Math.min(millis * 2, maxSleep) * (.9 + Math.random() / 5));
   }
 
-  public static List<KeyValue> scan(ClientContext context, ScanState scanState, long timeOut) throws ScanTimedOutException, AccumuloException,
+  public static List<KeyValue> scan(ClientContext context, ScanState scanState, int timeOut) throws ScanTimedOutException, AccumuloException,
       AccumuloSecurityException, TableNotFoundException {
     TabletLocation loc = null;
     Instance instance = context.getInstance();
